chore: extract ffow into multiple files
